# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "AI Gateway API",
    "description": "Central AI Gateway backend service"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ]
          },
          "db": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ]
          },
          "redis": {
            "type": "string",
            "enum": [
              "ok",
              "error"
            ]
          }
        },
        "required": [
          "status",
          "db",
          "redis"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string",
                "nullable": true
              },
              "details": {
                "type": "object",
                "nullable": true,
                "additionalProperties": {
                  "nullable": true
                }
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
      "RegisterInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "full_name": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "full_name": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "created_at"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_in": {
            "type": "number"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in"
        ]
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "auth_type": {
            "type": "string",
            "enum": [
              "jwt",
              "api_key"
            ]
          }
        },
        "required": [
          "user_id",
          "email",
          "role",
          "auth_type"
        ]
      },
      "CreateApiKeyInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "expires_in_days": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          }
        },
        "required": [
          "name"
        ]
      },
      "ApiKeyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "key_prefix": {
            "type": "string"
          },
          "rate_limit_per_min": {
            "type": "number"
          },
          "expires_at": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "warning": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "key",
          "key_prefix",
          "rate_limit_per_min",
          "expires_at",
          "created_at",
          "warning"
        ]
      },
      "ApiKeyListResponse": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid"
            },
            "name": {
              "type": "string"
            },
            "key_prefix": {
              "type": "string"
            },
            "rate_limit_per_min": {
              "type": "number"
            },
            "last_used_at": {
              "type": "string",
              "nullable": true
            },
            "expires_at": {
              "type": "string",
              "nullable": true
            },
            "revoked_at": {
              "type": "string",
              "nullable": true
            },
            "created_at": {
              "type": "string"
            }
          },
          "required": [
            "id",
            "name",
            "key_prefix",
            "rate_limit_per_min",
            "last_used_at",
            "expires_at",
            "revoked_at",
            "created_at"
          ]
        }
      }
    },
    "parameters": {}
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Checks connectivity to PostgreSQL database and Redis cache.",
        "tags": [
          "System"
        ],
        "responses": {
          "200": {
            "description": "Service is fully healthy",
            "content": {
              "application/json": {
                "example": {
                  "status": "ok",
                  "db": "ok",
                  "redis": "ok"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok",
                        "error"
                      ]
                    },
                    "db": {
                      "type": "string",
                      "enum": [
                        "ok",
                        "error"
                      ]
                    },
                    "redis": {
                      "type": "string",
                      "enum": [
                        "ok",
                        "error"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "db",
                    "redis"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Service is degraded or unhealthy",
            "content": {
              "application/json": {
                "example": {
                  "status": "error",
                  "db": "error",
                  "redis": "ok"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok",
                        "error"
                      ]
                    },
                    "db": {
                      "type": "string",
                      "enum": [
                        "ok",
                        "error"
                      ]
                    },
                    "redis": {
                      "type": "string",
                      "enum": [
                        "ok",
                        "error"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "db",
                    "redis"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "full_name": {
                      "type": "string",
                      "nullable": true
                    },
                    "role": {
                      "type": "string"
                    },
                    "created_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "email",
                    "role",
                    "created_at"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already registered"
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "summary": "User login",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "expires_in": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "access_token",
                    "token_type",
                    "expires_in"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password"
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "summary": "Get current caller info",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current authenticated caller info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "role": {
                      "type": "string"
                    },
                    "auth_type": {
                      "type": "string",
                      "enum": [
                        "jwt",
                        "api_key"
                      ]
                    }
                  },
                  "required": [
                    "user_id",
                    "email",
                    "role",
                    "auth_type"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/v1/api-keys": {
      "post": {
        "summary": "Create an API key",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "key": {
                      "type": "string"
                    },
                    "key_prefix": {
                      "type": "string"
                    },
                    "rate_limit_per_min": {
                      "type": "number"
                    },
                    "expires_at": {
                      "type": "string",
                      "nullable": true
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "warning": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "key",
                    "key_prefix",
                    "rate_limit_per_min",
                    "expires_at",
                    "created_at",
                    "warning"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "API keys cannot manage API keys"
          }
        }
      },
      "get": {
        "summary": "List user API keys",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "key_prefix": {
                        "type": "string"
                      },
                      "rate_limit_per_min": {
                        "type": "number"
                      },
                      "last_used_at": {
                        "type": "string",
                        "nullable": true
                      },
                      "expires_at": {
                        "type": "string",
                        "nullable": true
                      },
                      "revoked_at": {
                        "type": "string",
                        "nullable": true
                      },
                      "created_at": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "key_prefix",
                      "rate_limit_per_min",
                      "last_used_at",
                      "expires_at",
                      "revoked_at",
                      "created_at"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "API keys cannot manage API keys"
          }
        }
      }
    },
    "/v1/api-keys/{id}": {
      "delete": {
        "summary": "Revoke an API key",
        "tags": [
          "API Keys"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "API key revoked"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "API keys cannot manage API keys"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    }
  }
}